# swagger_full.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "App Base API",
        "description": "API REST base com autenticação JWT, gestão de utilizadores, upload de ficheiros e sistema de refresh tokens. Desenvolvida para servir como ponto de partida em novos projectos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-parceiro-dexa.appteste.info"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
...
    "paths": {
...
        "/api/v1/leads/{publicId}": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Detalhes do lead",
                "description": "Retorna todos os dados de um lead, incluindo notas",
                "operationId": "buscarDetalhes",
                "parameters": [
                    {
                        "name": "publicId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeadResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lead não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Leads"
                ],
                "summary": "Editar lead",
                "description": "Actualiza os dados de um lead. Se o novo estado for CONVERTIDO é obrigatório enviar o comprovativo (caso ainda não exista). O comprovativo é opcional para os restantes estados.",
                "operationId": "editar_1",
                "parameters": [
                    {
                        "name": "publicId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "dados": {
                                        "$ref": "#/components/schemas/LeadRequest"
                                    },
                                    "comprovativo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "dados"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Lead actualizado com sucesso"
                    },
                    "400": {
                        "description": "Campos inválidos ou comprovativo em falta para CONVERTIDO",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lead ou pacote não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Leads"
                ],
                "summary": "Eliminar lead",
                "description": "Elimina permanentemente um lead e as suas notas",
                "operationId": "eliminar_1",
                "parameters": [
                    {
                        "name": "publicId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lead eliminado com sucesso"
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lead não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Listar os meus tickets",
                "description": "Retorna todos os tickets publicados pelo parceiro autenticado (paginado)",
                "operationId": "listar_1",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "tamanho",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de tickets",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseTicketResumoResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado — requer papel PARCEIRO",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Submeter um ticket",
                "description": "Cria um novo ticket de suporte com estado ABERTO",
                "operationId": "criar_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TicketRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Ticket criado com sucesso"
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado — requer papel PARCEIRO",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/leads": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Listar os meus leads",
                "description": "Lista os leads do parceiro autenticado com filtros opcionais por estado, data inicial e data final",
                "operationId": "listar_2",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "LEAD",
                                "PENDENTE",
                                "CONVERTIDO",
                                "PERDIDO"
                            ]
                        }
                    },
                    {
                        "name": "dataInicial",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "dataFinal",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "tamanho",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de leads",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseLeadResumoResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Criar lead",
                "description": "Cria um novo lead associado a um parceiro e pacote",
                "operationId": "criar_2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LeadRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Lead criado — devolve o publicId",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Campos inválidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado — requer papel PARCEIRO",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parceiro ou pacote não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/leads/overview": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Overview dos meus leads",
                "description": "Retorna o total de leads e o total de faturação do parceiro autenticado para uso em dashboard",
                "operationId": "overview",
                "responses": {
                    "200": {
                        "description": "Overview calculado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeadOverviewResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StandardError": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "msg": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "LeadRequest": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string"
                    },
                    "telemovel": {
                        "type": "string"
                    },
                    "nif": {
                        "type": "string"
                    },
                    "pacotePublicId": {
                        "type": "string",
                        "minLength": 1
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "LEAD",
                            "PENDENTE",
                            "CONVERTIDO",
                            "PERDIDO"
                        ]
                    }
                },
                "required": [
                    "nome",
                    "pacotePublicId"
                ]
            },
...
            "TicketRequest": {
                "type": "object",
                "properties": {
                    "conteudo": {
                        "type": "string",
                        "minLength": 1
                    },
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "CANCELAR",
                            "BUG",
                            "RECLAMACAO",
                            "OUTRO"
                        ]
                    }
                },
                "required": [
                    "conteudo",
                    "tipo"
                ]
            },
...
            "LeadResponse": {
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "telemovel": {
                        "type": "string"
                    },
                    "nif": {
                        "type": "string"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "LEAD",
                            "PENDENTE",
                            "CONVERTIDO",
                            "PERDIDO"
                        ]
                    },
                    "comprovantivoUrl": {
                        "type": "string"
                    },
                    "pacotePublicId": {
                        "type": "string"
                    },
                    "pacoteNome": {
                        "type": "string"
                    },
                    "pacotePreco": {
                        "type": "number"
                    },
                    "parceiroPublicId": {
                        "type": "string"
                    },
                    "parceiroNome": {
                        "type": "string"
                    },
                    "usuarioPublicId": {
                        "type": "string"
                    },
                    "usuarioNome": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "notas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NotaLeadResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer Authentication": {
                "type": "http",
                "name": "Bearer Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}